Deploy: Add basic pre-deploy merge logic
Summary:
In order to make deploying faster, we want to do a lot of the work that
is involved in a deploy ahead of time: things like making genfiles,
testing, uploading, deploying to a non-default. We can, for example, do
this on branches in the queue, and for commonly-deployed branches.

Chances are, we are going to be doing this while another deploy is
happening, which is not in master. So the basic logic is that we will
kick off two pre-deploy-processing jobs: one for the branch we're trying
to deploy, merged with the current deploying branch + latest
translations, and one with the current deploying branch + master +
latest translations (in case the current deploy fails).

Test Plan: Mostly prayer-based at the moment

Reviewers: csilvers, benkraft

Subscribers: drosile

Differential Revision: https://phabricator.khanacademy.org/D40787