Try again to move to `git checkout -f`.
This time we don't use the `--recurse-submodules` flag, because it
resets `master` on the submodule, not just `HEAD`.  Instead, we call
`git submodule update` manually, which does the right thing.  Slower,
but not too bad since we don't have to do `--init` or `sync` first (I
don't think).

Another speedup is to use `git status` output to say what submodules
need to be updated, instead of using `git submodule foreach` to update
all of them.  That should be much faster in the common case.

Test plan:
./safe_git_test.sh