Groovy script to merge a branch with master and run all webapp tests on it.
Summary:
As a part of the TSM project "Making Tests Faster", we want to run a job that runs all
webapp tests when pushed to github. The github 'push' notifications will come through
the Deploy State Store (coordinating with the "Make Deploys Faster" team) later on.

We will use this change to manually trigger a "clone-webapp-test" jenkins job. It will send
messages with test updates to the #make-tests-faster slack channel.

Test Plan:
Ran `groovy jobs/clone-webapp-test.groovy`
Hoping to deploy it and create a job following https://jenkins.khanacademy.org/job/groovy-template/ to test it out.

Reviewers: johnsullivan, benkraft, jacqueline

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40766