Explicitly keep version toc-files when cleaning ka-static
Summary:
The comment describes the problem -- when we copy the manifest file, we
don't update its reference to itself, nor its reference to the (also
copied) top-level manifest (`genfiles/manifests/toc-*`).  This might
cause us to accidentally delete that file even perhaps for the live
version (if it was a dynamic-only deploy), which is super scary!  (I'm
honestly amazed this didn't break the site; it did break the subsequent
deploy.)

So now we just explicitly keep the corresponding toc files.  This is a
bit fragile; I've added a TODO for a better way (and will add the same
TODO to webapp) but I'm not sure it's worthwhile.

Test Plan: cross fingers

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40818