Add new job to merge branches (WIP)
Summary:
Now that the buildmaster is tracking a lot more of the state of the
deploy, we also want it to be in charge of merging branches -- then
everything else Jenkins does will apply to commits.  But of course
Jenkins still has to do the merge!  This builds the job to do that.  It
still has some stopships, which depend on using code from D41211 to talk
to the buildmaster.

One of the exciting things this opens up is that we no longer have to do
so much work around submodules.  Since we don't need to end up with a
valid checkout, just one that can do the merge, we need to do very
little.  (Git will ignore or clobber any problems in the local tree on
merge, and any merge in a submodule would have to be done by the user
anyway.)  This should be a lot faster -- and then really only the test
and build jobs will need to mess about with submodules, and only once
since they already have a finalized commit.

Eventually, we will also be able to clean up a bunch of logic in those
jobs.  But that will come later once we've transitioned over to the new
system.

Test Plan: `groovy jobs/merge-branches.groovy`; cross fingers

Reviewers: csilvers, drosile, sergei, joshuan

Differential Revision: https://phabricator.khanacademy.org/D41231