Generalize version deletion job
Summary:
The buildmaster wants to be able to just say "delete this version" and
have jenkins do it.  Now it will!  I stole all the interesting parts
from the delete-znd job, and basically just removed the check that it's
a znd.  The other job now calls out to the new one after checking we
have a znd (which still seems like a good idea for `sun: delete znd`).

Test Plan: crossed fingers

Reviewers: drosile, sergei

Reviewed By: drosile

Differential Revision: https://phabricator.khanacademy.org/D41280