Don't notify buildmaster unless processing 1 git revision
Previously, we were relying on buildmaster to know whether it was
receiving a single git-sha (that it knows how to deal with) or a branch
name (which it does not).  This diff moves that logic to Jenkins, so
buildmaster can safely assume it will receive a single git-sha.

Auditors: benkraft

Test Plan:
Push, test manually on jenkins