Fix arg to quickClone
I had typoed it.  But fixing it made me realize that passing in
`master + foo` wouldn't work!  We want to pass in the first branch.
This makes the checkout a no-op if we are making a fresh clone, but
it should be a fast no-op, and passing the right data around would be a
little annoying -- as is only `safe_git.sh` knows if it created the
directory or not.

Test plan: deploy, run the job

Auditors: csilvers, sergei