Buildmaster response >=400 should fail jenkins job
Summary:
If communicating with buildmaster results in a status >=400, then
something went wrong, and the build should be marked as being failed.
What to do once we're in this state is still TODO(buildmaster).

Even aborted jobs will be marked as failed if telling the buildmaster
about the abort fails.

Test Plan:
 - make sure buildmaster requests will fail
 - run this jenkins job using the 'Replay' feature
 - The job succeeds|aborts, but then is marked as failed
 - appropriate FAILED slack notification

Reviewers: csilvers, benkraft

Reviewed By: csilvers, benkraft

Subscribers: drosile

Differential Revision: https://phabricator.khanacademy.org/D41383