Catch errors trying to resolveCommitish(br1+br2)
We were using resolveCommitish to check whether GIT_REVISION was
actually a single sha (which it generally will be once buildmaster is
doing all our merging for us), or if it's a branch-name or multiple
branchnames in br1+br2+.. format.

Turns out resolveCommitish will fail the job if it can't resolve the
commit, which happens since br1+br2 is seldom a valid commit.

Let's just catch that error and interpret it to mean that GIT_REVISION
is not a single sha.

Auditors: benkraft

Test Plan:
run webapp-test-new(master+sat) and it doesn't fail right away now!