Rename rune2etests.py to runsmoketests.py
Summary:
In D41436 we renamed the test runner for our e2e tests to be
runsmoketests.py. This change will need to land at approximately the
same time to ensure that the e2e test jobs continues to work as
expected.

Test Plan: Fingers crossed? Will ensure that the test runs after landing

Reviewers: csilvers, aasmund

Subscribers: