Add e2e-test-new job for Buildmaster integration
Summary:
This is a copy-paste of e2e-test, with a few tweaks to report to buildmaster.

When we are ready to promote this, we can rename it to replace
`e2e-test`.

This is an initial instance of the e2e job. I think we'll need to make some
further changes before we're ready to use this. Namely:
- run e2e tests directly from buildmaster, after a build is completed
- don't run e2e tests as part of the deploy job in jenkins
- store DEPLOY_URL in buildmaster somehow so it can successfully call this job

Test Plan: run the job manually and see if it reports to #bot-testing

Reviewers: benkraft, sergei

Subscribers: