Give each test-worker 2 hours to run, not 1
When running make-allcheck, tests can take a long time to run because
we have a lot of them (especially since make-test-db-tet re-runs most
tests again).  So up the timeout.

Better would be to have the timeout be a param, so regular make-check
will time out sooner than make-allcheck.  Or maybe we have a
multiplication-factor that we add in based on maximum test-size.  In
any case, a problem for another day.

Auditors: carter

Test plan:
Fingers crossed for next make-allcheck run