Create a deploy tag rather than a deploy branch.
Summary:
The thing-we-deploy is better represented as a tag than a branch,
semantically, since it's not something we develop on.  And phabriactor
starts to slow down when we have too many branches, so there's a
practical reason to make this a tag as well.

Test Plan:
Fingers crossed.  I at least ran
    groovy jobs/deploy-webapp.groovy
with no unexpected errors.  But we'll have to see how deploys go once
this is landed!

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D41798