Turn old branches into tags.
Summary:
Add a new cleanup task that turns branches that haven't been worked on
a for a while -- like, 6 months -- into tags.  No content is deleted,
but if you ever wanted to continue working on that branch again you'd
need to recreate it in git

The reason we bother is that phabricator daemons do `git branch
--contains` which gets very slow when there are a lot of branches.

Test Plan:
I ran the command by hand, though on my local machine and not on
jenkins.  So somewhat fingers crossed.

Reviewers: colin, amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D41799