Add options for deploy-webapp to just build, used by the buildmaster
The buildmaster wants to do its own gluing-together of the build step
(i.e. uploading to GAE/GCS) with running tests and eventually promoting
the new build to default.  (It will then be composing these in more
complex ways, often doing the build and test steps while the prior
promote is happening.)  So it needs to be able to ask Jenkins to do the
build alone.  This creates that job, currently in the form of a flag to
deploy-webapp, which causes us to only do the build, and to phone back
home to the buildmaster.

I did this as a flag rather than as a separate job because a lot of the
code is the same.  Eventually, I'll be adding a similar flag for the
promote-only part, and then, once the buildmaster is live and we no
longer need the option to do both, I'll look at splitting them into
separate jobs.  This way we never have to maintain two versions of the
same logic.

Other than that, this came out reasonably simple, I think.  There will
be a follow-up diff to add an explicit BASE_REVISION to use in deciding
what deploys to do, similar to the flag to webapp-test.  (And, of
course, one to allow the buildmaster to actually run this job!)

Test plan:
crossed fingers; will test having the buildmaster run this
job once it knows how.

Reviewers: csilvers, drosile