Add a BASE_REVISION option to deploy-webapp
Summary:
This is the base against which we run should_deploy, to decide (if
unspecified) whether to deploy to static, dynamic, both, or neither.  We
will now need to be able to specify it explicitly for builds, because
they may run before the previous deploy is done.

For example, if version A (a static+dynamic deploy) is currently
deploying, and we want to speculatively build version B (which only
changes static files), currently when trying to build A+B we would think
we needed a dynamic deploy.  But in reality, once we are setting default
on B, A will already be deployed, so we will only need the static
deploy.  So we'd set the BASE_REVISION to A.

(It's not clear what setting this to anything other than the currently
live version would mean for the promote step, so I ignored it if we're
doing that, to prevent mistakes.)

Depends on D42093.  Other than that, all the hooks we needed already
exist; this was just a matter of setting them.

Test Plan:
crossed fingers; will test once it's hooked up to the
buildmaster.  (And will test a normal deploy when I push this to make
sure I didn't break anything.)

Reviewers: drosile, csilvers

Reviewed By: drosile, csilvers

Differential Revision: https://phabricator.khanacademy.org/D42095