Give `git submodule update` a bit more time.
I thought if you called `update` without calling `init` first, then
nothing would happen, but apparently it tries to check out everything
anyway?  At least sometimes.  Perhaps it's that the make-allcheck repo
was in a weird state.  In any case, this `git submodule update` was
trying to clone a bunch of repos and it wasn't finishing in a minute,
so make-allcheck was failing.

It looks like it took 2.5 minutes in the end, so upping to 10 minutes,
to match other update calls, should be fine.

While in the area, I also changed an argument to syncTo to match the
API better.  There is no functional change.

Auditors: joshuan

Test Plan:
Fingers crossed