Fix a safe-checkout problem discovered in real life.
I didn't do enough tests where files changed on the remote *and*
changed locally.  This was causing a problem that I had to solve with
another `--force` flag.

Sadly, the new `-f` flag adds a second or two to the runtime of
destructive-checkout, because it actually does two things:
1) Does a checkout even if the submodule is already at the right commit
2) Does a `git checkout -f` when it does that checkout

We want (2) but not (1).  Sadly there's no way to separate them that I
can see.

Auditors: joshuan

Test Plan:
./safe_git_test.sh

I verified it failed before the addition of `-f`, and passed after.