Make deploy-webapp a wrapper job to allow concurrent builds
Summary:
The buildmaster uses a new mode of deploy-webapp, with STAGES=build, to
do only a build of webapp, and not set it as default.  We'd like to be
able to run this new version in parallel with itself and with the old
flow -- since it's not doing a full deploy there's no problem running it
in parallel.

To accomplish that, this commit makes deploy-webapp into a wrapper
script that just passes through its params to a new
`deploy-webapp-core`, which contains all the logic and allows concurrent
builds.  Thus the buildmaster can just call the underlying job, and
avoid the concurrency restriction.  This also means we can remove the
potentially-confusing buildmaster-only options from the job people might
start manually.

This is all a bit ugly, sadly, but our use of globals makes it hard to
make two completely separate jobs, so we have to share.

Test Plan:
deploy it, then while a deploy is going, `sun2: build <branch>` should
start doing work immediately.

Reviewers: drosile, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D42132