Add a new "promote" option for STAGES
This is the counterpart to D42076, to add a "promote-only" option to
deploy-webapp, which assumes everything is built already, and we just
need to flip it over to default.  This will be run by the buildmaster.

This one came out pretty easy, since all the scaffolding was already
there!  I just had to add a few more conditionals.  In a few cases we
may want to change exactly how the slack messages work in the future,
but I left that as a TODO.

Depends on D42132 (only to avoid a nasty merge).

Test plan: crossed fingers, until I set up the buildmaster integration

Reviewers: drosile, csilvers