Allow DEPLOYER_USERNAME to start with <@ as well
The full slack at-mention syntax these days is `<@Uxxxxxxx>`, which the
buildmaster passes in so it can include multiple mentions.  If we get
such syntax, we should just leave it be -- eventually it should
probably be the preferred form.

Test plan:
deploy with the buildmaster, don't see an extra `@` in slack messages.

Auditors: csilvers