Add highmem-instance option for znd-deploys
Summary:
This commit is part of https://khanacademy.atlassian.net/browse/INFRA-851.
Associated webapp commit: https://phabricator.khanacademy.org/D42780.

Znds often struggle to fill caches and hit time out deadlines making the znd unusable. This diff adds a new highmem-instance param which allows us to switch the default instance-class used for znds to B8 in the deploy/deploy_to_gae script.

Highmem-instance defaults to true for znd-deploys, but can (and should be) explicitly set to false in the case that high traffic is expected on that znd as these instances are significantly more expensive.

This also updates the alertlib submodule as that was accidentally reverted in a previous commit.

Test Plan:
1. Following deploy of associated webapp diff and this commit, run a deploy-znd job.
2. Verify that znd version configs in google cloud console have instance_class set to B8.

Reviewers: benkraft, csilvers

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D42781