Changes to deploy slack messages when using the buildmaster
This makes several changes to deploy jobs to make things clearer when
using the buildmaster.

First, we remove the start-deploy message for jobs with STAGES, since
there are lots of build jobs floating about and it gets confusing.
Similarly, we use the option added in webapp 600080bb to suppress the
changelog at the start of deploy, and send it along with the set-default
message instead.

Second, for all the deploy jobs, we allow passing a parameter with a
description of the branch, which if specified is used in slack messages
instead of the GIT_REVISION; this is useful because the buildmaster
generally passes an explicit commit.

Finally, I removed success-path notifications from a few things that now
happen without user action (deploy-webapp-core start, and delete-version
end).  These are no longer useful unless something goes wrong with the
buildmaster itself, which is uncommon and there are better ways to
debug.  (For example, it pretty consistently reports crashes back to
slack!)

One day, I hope to move all of this messaging to the buildmaster which
can be a lot smarter about exactly what it says when.  But that day is
not today.

Test plan: crossed fingers; will also run a test deploy

Auditors: csilvers, drosile