Use safeSyncToOrigin in most places.
Summary:
safeSyncTo is meant to be used in situations where we have committed
(but not yet pushed) local changes.  That doesn't happen in groovy
scripts proper (which don't update git state hopefully!)

So I got rid of safeSyncTo as a wrapper in kaGit, and now only expose
safeSyncToOrigin, which is what groovy scripts want: reset the branch
locally to be the same as what's on origin.

Fixes https://khanacademy.atlassian.net/browse/INFRA-275

Test Plan:
Fingers crossed -- we'll see what happens to jenkins jobs after this
lands!

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: joshuan

Differential Revision: https://phabricator.khanacademy.org/D43021