Do not send to slack on ABORT for a bunch of "downstream" jobs.
Some jobs can be called by other jobs.  When the "main" job fails, it
will abort any of the "downstream" jobs that are running.  If those
downstream jobs are set to notify slack on abort, they'll do so, which
tends to be confusing than edifying.  As a concrete example, when a
deploy fails while tests are running, deploy-webapp will cancel
webapp-test, which will send an abort message to slack, causing the
deployer to think the problem is a failed test when it's not.

So I went through all the jobs and used my domain knowledge to
determine which jobs were "user facing" and which were "downstream",
and turned off slack alerting on ABORT for the latter.  I kept up the
alera alerting for ABORT in these cases, so there's still some record.

Auditors: benkraft

Test Plan:
`groovy foo.groovy` for all modified files.