Say "uploaded" instead of "deployed", in the message after we upload
In the past, this was right before you would set default, so it made
plenty of sense, but now it's potentially earlier, and that's confusing
and may make you think it has done something very wrong!  Now we say
"uploaded" to be clearer.

Test plan: read the message

Auditors: csilvers

Subscribers: cora