Split up deploy-webapp-core into build-webapp and deploy-webapp
Summary:
The big day has come!  Since D43199 we are no longer using the
STAGES=all mode that does both build and deploy in the same job.  Now,
instead, we can split up build and deploy into separate jobs.  This
commit does that.  I haven't yet removed deploy-webapp-core -- I'll do
that in a separate commit, after I point the buildmaster to the new
jobs.

The split was mostly pretty clean.  There is still some duplication --
the merges at the beginning, and `_alert` -- but I think that's ok for
now, especially since our use of globals makes it hard to share code.
(Although suggestions are welcome!)  It may diverge a bit more soon,
anyway; and if not we can separate some of it out.  Relatedly, I did not
yet make any attempt to make nontrivial behavior simplifications (e.g.
simplifying the git tags) at this time; that will be in a separate
commit to keep things simple.  This commit simply splits the existing
code up exactly as it stands.

Test Plan: deploy with it!  And cross fingers.

Reviewers: csilvers, drosile, jacqueline

Differential Revision: https://phabricator.khanacademy.org/D43208