Clean up build-webapp notifications
This commit makes two changes to the build-webapp notifications.  First,
we no longer notify on abort.  This is basically for the same reason
we've turned it off for tests and such; it almost always means something
*else* happened -- either a user action or a failure -- and that's the
relevant part, not whatever it caused to abort.  Second, we no longer
at-mention in our notifications, because the buildmaster will do so when
relevant.

Test plan: crossed fingers; deploy it

Reviewers: csilvers, drosile, jacqueline