Change weekly-maintenance to call the jobs one at a time from within groovy.
Summary:
It's now not quite such a monolithic shell script anymore.  This will
give us insight into how long each of the maintenance jobs takes, and
will let us continue to other jobs if a particular job fails.

The first part was changing the shell script along the lines of what
we do for safe_git_test.sh.  We now run each job in its own subshell,
and keep track of which ones failed.

I also automated the collection of jobs to run, so you don't need to
both define a new function *and* list it at the bottom of the file.

The second part was changing the jenkins job to take advantage of this
new functionality.  It now lists which jobs to run and runs them just
one at a time.  We do each one in a separate jenkins stage.

Test Plan:
Ran `./weekly-maintenance.sh -l` without error.  I also started to run
   ./weekly-maintenance.sh
and saw it start doing the docker-cleanup before I got scared and hit
control-C.

I also ran
   groovy jobs/webapp-maintenance.groovy

Reviewers: amos, colin

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D43267