Phone home to the buildmaster at start-of-job
This un-reverts and replaces eb025733.  In addition to the changes from
there, I had to change where we decide whether to send to the
buildmaster -- we were now trying to do that before the relevant globals
were set up.  Now we just pass in the input GIT_REVISION directly, and
the buildmaster just exits immediately if it's not a sha.  I hope to
clean this up even further -- to say your merges always have to go
through the buildmaster -- soon.

Test plan: push, tests should continue working

Auditors: csilvers