Only try to prune bad refs-to-remotes in clean_invalid_branches().
With the new phabricator diff-tagging, we have tens of thousands of
tags in our branches.  That's too much for `clean_invalid_branches()`
to go through.  I don't think we have much risk of tags becoming
invalid anyway; the problem seems to be only for remotes (when a
remote branch is deleted on github, perhaps).  So I'll only check for
those.

Auditors: amos, colin

Test plan:
Fingers crossed, mostly.