Use the "alternates" file instead of git-new-workdir.
Summary:
We've been having some trouble with new-workdir with failing to
acquire locks properly when updating submodules, leading to conflicts
that cause git operations to fail with "Can't acquire lock"
(presumably because some other process has it).

Let's solve this problem once and for all by moving away from the
hacky new-workdir methodology to the somewhat less hacky
"alternates-dir" methodology:
    http://dustin.sallings.org/2008/12/30/git-alternates.html
    https://mirrors.edge.kernel.org/pub/software/scm/git/docs/gitrepository-layout.html

We didn't do this before because the alternates dir doesn't work for
submodules (I don't think), but now we're only using new-workdir for
the top-level webapp module anyway, so I think life should be fine
now.

(The other alternative, which is perhaps even better suited for what
we're doing, is `git worktree`, but the man page says "it's beta and
doesn't work well with submodules", so I'm not using it for now.)

The way 'alternates' works is that when you clone or gc, it looks to
see if the objects you need are available in the "alternates"
location, and if so it doesn't bother to save it locally.  This saves
a bunch of space since we only need to have the objects in one place
(the place mentioned in "alternates").

However, each repo still downloads new objects on "git fetch", which
takes up space.  So we now run, once a week, "git gc" in every repo,
to make it switch to using the version of the object in "alternates"
instead.

This is not as ideal as the old approach, since we use more and more
space over the week before reclaiming it over the weekend.  But it
should be safer.  It also speeds up the git setup operation on every
jenkins job a bit, so that's good.

Should be the last part of fixing
https://khanacademy.atlassian.net/browse/INFRA-1051

Test Plan: Fingers way crossed.  I'll be looking at jenkins jobs after these go.

Reviewers: benkraft, jacqueline

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D43841