Replace a `&&` with an `if`.
We were seeing this behavior:
    $ git status --porcelain -z | tr '\0' '\012' | cut -b4- | while read f; do [ -d "$f" ] && echo $f; done
    Saving bigfile: 37e8cdc65b6502257c6c3892897f55925bdaf62c
    Saving bigfile: d81197cc4f719a0839e698f8a6fb1efe4865fa6f
    $ echo $?
    1

My understanding is that if you do a `&&` which fails, it does not
cause the script to fail (due to the `set -e`).  But apparently it
is.  Hopefully changing from `&&` to `if` will fix that problem.

This is somewhat a shot in the dark.

Auditors: benkraft

Test plan:
Fingers crossed