Allow passing slack threads in jenkins jobs
Summary:
We'd like to move most of the messages in a deploy before the
set-default stage into a thread, to keep the main deploys channel a bit
quieter.  This commit adds support for doing so: the merge, build, test,
and end-to-end test jobs all now take a SLACK_THREAD parameter which
will send to that thread.  (Build also now supports SLACK_CHANNEL for
consistency; deploy remains only in the main `#1s-and-0s-deploys`.)

This was pretty straightforward, I just had to do it in a bunch of
different places.  The code for tests is a little special: since
sometimes tests run on old branches I made sure we're
backwards-compatible with code before a26a7c105.  I didn't do this for
builds since we always merge to master for those, so this depends on
D43902.

None of this will happen automatically just yet; keep an eye out for a
separate diff in the buildmaster which will set this flag.

Test Plan:
- `groovy jobs/<job>.groovy` gives only the usual errors, for each
  changed job
- crossed fingers
- will deploy, setting the flag manually as a test.

Reviewers: csilvers, jacqueline

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D43904