Refactor webapp-test and e2e-test to hold workers continuously
Summary:
Currently, webapp-test and e2e-test spawn a bunch of workers to do
setup (while determining test-splits), wait for that to finish, and then
spawn a bunch of workers again to do the tests.  Sadly, we may not get
the same workers, and we have to wait for them twice (and wait for all
the setups to finish before we can start the second wait), so this is
less than effective and kinda slow.

Now, instead, we grab the workers right away, have them start setup, and
then wait for test-splitting to finish before proceeding to unstash and
run the tests.  This should, if it works correctly, have all the same
benefits, and run a lot faster.

This should make no logical changes: it just refactors the control flow
to something that should be equivalent.

Test Plan:
- `groovy jobs/*-test.groovy` gives just the usual errors
- fingers super crossed

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D43975