Refactor onWorker wrappers to share code
Summary:
Now that we have 4 kinds of workers -- and I'm about to add another --
it's getting sad to have to copy all the code.  Literally the only
difference between `onTestWorker` and `onBuildWoerk` was the label.  So
now I just created a single onWorker, and you pass it the label.  I
documented the list of labels; it seemed clearer than writing a wrapper
for each kind.

The two jobs that were not using the standard code, build-current-sqlite
and page-weight-test, had slightly different logic.  In the case of
build-current-sqlite, the only difference is that we now use the
webapp-workspace, instead of some default directory.  page-weight-test
catches errors in a slightly different place, which I believe is
equivalent.

In a separate commit, I'll clean up notify to make more sense in the
current context.

Test Plan:
- `groovy vars/*.groovy jobs/*.groovy``
- cross fingers

Reviewers: csilvers, jacqueline, joshuan

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D43992