Build znds on build workers
I've been wanting to do this for a while so we can do them in parallel,
but the more immediate need now is CPU: when a deploy-znd is running it
uses up all the CPU on the master, and everything else grinds to a
crawl.  Let's just solve both problems by running it on a worker.  I
used a separate workspace, so znds don't make a mess of our build caches
for the main deploy.

Test plan:
- `groovy jobs/deploy-znd.groovy jobs/build-webapp.groovy vars/onBuildMaster.groovy`
- cross fingers

Reviewers: csilvers, jacqueline