Build znds on build workers
Summary:
I've been wanting to do this for a while so we can do them in parallel,
but the more immediate need now is CPU: when a deploy-znd is running it
uses up all the CPU on the master, and everything else grinds to a
crawl.  Let's just solve both problems by running it on a worker.  I
made a separate worker type (depends on D43992), so znds don't make a
mess of our build caches for the main deploy.  (It was easier than using
a separate workspace -- this replaces D43984.)

Test Plan:
- `groovy jobs/deploy-znd.groovy`
- cross fingers
- deploy a znd

Reviewers: csilvers, jacqueline

Differential Revision: https://phabricator.khanacademy.org/D43993