Drive the tests from a test worker
Now that we're running tests on every commit, it's not unheard of to
have a dozen tests running at once.  That means a bunch of tests
starting at once too, which means a bunch of tests doing a bunch of git
ops on the master.  Now that we have plenty of workers, there's no need:
we can just drive the tests from a worker too, and have it farm out to
even more workers as necessary.  This should hopefully keep things
moving smoothly, and prevent the master from getting too overwhelmed.

Test plan: `groovy jobs/*.groovy vars/*.groovy`; fingers very crossed

Reviewers: csilvers, jacqueline

Subscribers: kai