Remove singleton on webapp-test
The buildmaster now handles this!  In fact it's even smarter: if two
people ask for tests on the same commit, it'll just run them once.  (It
does have some UI issues, in that it doesn't always tell you very
clearly what happened, but so did singleton!  I've created INFRA-1118 to
track.)  So we remove the Jenkins version, which will also allow moving
tests to workers.

Note that content-publish-e2e-test does still use singleton, so I didn't
get rid of it just yet.  But we may want to: that job has been failing
for a while and we haven't really cared.

Test plan: deploy it, cross fingers

Auditors: csilvers