Change Jenkins jobs to think in terms of deploying services
Summary:
This is the corresponding change to D44432, for our groovy scripts.
Now, instead of DEPLOY, you set SERVICES, which is a comma-separated
list of services.  Currently those are just "static" and "dynamic", but
this will make it easier to add more.  The logic is as before: I expect
a few of these checks will change as we add more services (for example
static copying will only be needed if we are deploying dynamic but not
static), but we'll get there when we get there.

At this time I did not change deploy-znd, because I don't expect to add
znd support for other services at this time.

Depends on D44432, for the new syntax to should_deploy.py.

Test Plan:
- `groovy jobs/build-webapp.groovy ; groovy jobs/deploy-webapp.groovy`
  gives only the usual errors
- deploy it, and cross fingers

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: colin, #devops

Differential Revision: https://phabricator.khanacademy.org/D44434