Send merge to master error from the right dir, and update the message
Summary:
We need the right directory so we can send a message.  This also ensures
we catch failures making the git tag.  I also updated the message to
just tell you to ping dev-support, and added an entry to the dev-support
troubleshooting guide ("Advanced Troubleshooting" in
https://docs.google.com/document/d/1Zr0wwzbvPkmN_BFAsrMPZhccIb0HHJUgmLZfBUWYFvA/edit#).

Test Plan: crossed fingers

Reviewers: csilvers, jacqueline

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D44479