Don't run the ordering script twice in download-translations
Summary: When we run a download-translations job, if we don't specify any
override langs, we run a script to determine which language is the highest
priority and update that language. Unfortunately, that script is very slow (~8
minutes to run!) and we were running it *twice* every time we ran a job. This
changes that to only run the script once.

Test Plan:
 - Cross fingers?

Reviewers: tom, aasmund