Remove logic to re-use a passed date/username in deploy-znd
Now we just complain if you include the "znd" and add the prefixes
unconditionally.  I also refactored so we compute the version once,
avoiding any chance that we compute it differently in different places,
which I suspect may have been the source of bugs in the past.

Test plan: crossed fingers

Reviewers: csilvers, jacqueline