Don't add a tag if something already points to the commit.
Summary:
We need to add a tag when merging branches so that github doesn't
prune the commit as being unreachable.  But if something already
points to the commit -- a tag or a branch -- then it's reachable and
github won't prune it.  So we don't have to tag in that case.

Normally we wouldn't care about these extra tags, but they seem to be
confusing phabricator into sending out an audit email every time we
run a test.  We're hoping this will fix that problem.

Test Plan: groovy jobs/merge-branches.groovy

Reviewers: benkraft, jacqueline

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D45407