Refactor a little more of build-webapp to be more service-friendly
This removes a bunch of old version parameters for specific services.
Now we just have NEW_VERSION, which is the version for whichever
services will be deployed, BASE_REVISION_VERSION, which is the version
for whatever we're building off of, and DEPLOY_URL.  To keep them short
and simplify things, the Slack messages now say a list of services and
the new version, rather than the new entire version-tag.

Note that I will still be keeping the full version tags where they are
used in set-default, importantly in whats-happening (e.g. if someone
needs to find what version to roll back to), this is just for the
(threaded) messages to deployers about their builds.

Test plan: `groovy jobs/build-webapp.groovy`, deploy it

Reviewers: csilvers, colin