Add mark-conversions to build-webapp
Summary:
Once I had all the groundwork in place, this was surprisingly easy!
There were a few places I had to modify exactly how we check what to
do in a case or two -- in particular to fit with D45667.  (This makes
DEPLOY_URL super simple!)  Otherwise, it was just a matter of adding
another branch.

Note that we will not yet build mark-conversions automatically; it has
to be explicitly requested.  After I've added it to deploy-webapp, and
tested all the way through, I'll make a separate diff to
should_deploy.py to add it to the services we consider automatically
deploying.

Test Plan:
`groovy build-webapp.groovy` for syntax, but mostly, cross fingers,
deploy it, and test some deploys setting `SERVICES` to contain
`mark-conversions`

Reviewers: csilvers, colin

Reviewed By: csilvers, colin

Subscribers: #devops

Differential Revision: https://phabricator.khanacademy.org/D45673