Run khanalytics-test on a test worker
Summary:

Khanalytics tests are quick and previously we've just run them on the
master node. However, now we need to lint kotlin code, and the kotlin
linter is incompatible with the java version on the master node.

We've already fixed this on the workers, so this commit moves things to
a worker instead, which seems much safer than trying to upgrade the
version of java and risking breaking jenkins.

Test Plan:
- try to deploy again, hope

Auditors: benkraft