Fix deploy URL for tools-only versions
I knew I had simplified too much!  With our new static-version URL
scheme (D45667), we don't need special URLs for static-only deploys.
But the implementation means that URLs for a totally invalid version now
fail, whereas previously they went to the default version.  It turns out
we were relying on this for tools-only versions: they sent end-to-end
tests to `<version>-dot-khan-academy.appspot.com` even though that
version doesn't exist.

Now, instead, we fall back to choosing the version we are basing our
deploy on -- that is, either the BASE_REVISION_VERSION, if it exists, or
default.  This way, we test the new version of the e2e-test script, but
with our current live version of everything else.

Note: because I want to ship this fix right away, I pulled in the change
from D45673 (which I have not yet landed) to remove the static-only
version URLs, to keep things simple.  This is fine: the only dependency
is D45667 which already shipped.

Test plan: push, watch the next tools-only deploy

Auditors: csilvers, yeva